Made events.py the first api module.
Moved everything from models and schemas regarding events (and eventsignups) to confirm.py and renamed this to events.py

Added new utility functions to safely extend domain and validator.
Had to modify auth a little to work together with new structure - owner/registered_methods/public_methods etc are now stored in the domain aswell.

Needed to move declaration of storage to schemas.py because of this (since the info does not need to be in the model anymore, but in the domain)
