+++ amivapi/models.py
@@ -17,7 +17,6 @@ from sqlalchemy import (
     Enum,
     Boolean
 )
-from sqlalchemy.ext.hybrid import hybrid_property
 from sqlalchemy.orm import relationship, validates
 
 from amivapi.settings import PASSWORD_CONTEXT
@@ -204,97 +203,6 @@ class Session(Base):
     # user = relationship("User", foreign_keys=user_id, backref="sessions")
 
 
-class Event(Base):
-    __description__ = {
-        'general': "An Event is basically everything happening in the AMIV. "
-        "All time fields have the format YYYY-MM-DDThh:mmZ, e.g. "
-        "2014-12-20T11:50:06Z",
-        'methods': {
-            'GET': "You are always allowed, even without session, to view "
-            "AMIV-Events"
-        },
-        'fields': {
-            'price': 'Price of the event as Integer in Rappen.',
-            'additional_fields': "must be provided in form of a JSON-Schema. "
-            "You can add here fields you want to know from people signing up "
-            "going further than their email-address",
-            'allow_email_signup': "If False, only AMIV-Members can sign up "
-            "for this event",
-            'spots': "For no limit, set to '0'. If no signup required, set to "
-            "'-1'. Otherwise just provide an integer.",
-        }
-    }
-    __expose__ = True
-    __projected_fields__ = ['signups']
-
-    __public_methods__ = ['GET']
-
-    time_start = Column(DateTime)
-    time_end = Column(DateTime)
-    location = Column(Unicode(50))
-    allow_email_signup = Column(Boolean, default=False, nullable=False)
-    price = Column(Integer)  # Price in Rappen
-    spots = Column(Integer, nullable=False)
-    time_register_start = Column(DateTime)
-    time_register_end = Column(DateTime)
-    additional_fields = Column(Text)
-    show_infoscreen = Column(Boolean, default=False)
-    show_website = Column(Boolean, default=False)
-    show_announce = Column(Boolean, default=False)
-
-    @hybrid_property
-    def signup_count(self):
-        return len(self.signups)
-
-    # Images
-    img_thumbnail = Column(CHAR(100))  # This will be modified in schemas.py!
-    img_banner = Column(CHAR(100))  # This will be modified in schemas.py!
-    img_poster = Column(CHAR(100))  # This will be modified in schemas.py!
-    img_infoscreen = Column(CHAR(100))  # This will be modified in schemas.py!
-
-    title_de = Column(UnicodeText)
-    title_en = Column(UnicodeText)
-    description_de = Column(UnicodeText)
-    description_en = Column(UnicodeText)
-    catchphrase_de = Column(UnicodeText)
-    catchphrase_en = Column(UnicodeText)
-
-    # relationships
-    signups = relationship("EventSignup", backref="event",
-                           cascade="all")
-
-
-class EventSignup(Base):
-    __description__ = {
-        'general': "You can signup here for an existing event inside of the "
-        "registration-window. External Users can only sign up to public "
-        "events.",
-        'fields': {
-            'additional_fields': "Data-schema depends on 'additional_fields' "
-            "from the mapped event. Please provide in json-format.",
-            'user_id': "To sign up as external user, set 'user_id' to '-1'",
-            'email': "For registered users, this is just a projection of your "
-            "general email-address. External users need to provide their email"
-            " here.",
-        }}
-    __expose__ = True
-    __projected_fields__ = ['event', 'user']
-
-    __owner__ = ['user_id']
-    __owner_methods__ = ['GET', 'PATCH', 'DELETE']
-    __registered_methods__ = ['POST']
-
-    event_id = Column(Integer, ForeignKey("events.id"), nullable=False)
-    user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
-    email = Column(CHAR(100), ForeignKey("users.email"))
-    additional_fields = Column(Text)
-
-    """for unregistered users"""
-    _email_unreg = Column(Unicode(100))
-    _token = Column(CHAR(20), unique=True, nullable=True)
-    _confirmed = Column(Boolean, default=False)
-
-
 class File(Base):
     """This is a file that belongs to a study document.
 
@@ -385,18 +293,3 @@ class Purchase(Base):
     timestamp = Column(DateTime)
     type = Column(Enum("beer", "kaffi"))
     slot = Column(Integer)
-
-#
-# Permissions for custom endpoints
-# These are not created in the database due to __abstract__ = True
-# utils.get_class_for_resource will find these
-#
-
-
-class Storage(Base):
-    __expose__ = False  # Don't create a schema
-    __abstract__ = True
-    __registered_methods__ = ['GET']
-    __description__ = {
-        'general': 'Endpoint to download files, get the URLs via /files'
-    }
