+++ amivapi/schemas.py
@@ -5,16 +5,21 @@
 
 from sqlalchemy.ext.declarative import DeclarativeMeta
 
-from amivapi import models
 from amivapi.utils import make_domain, EMAIL_REGEX
+from amivapi.db_utils import Base
 
+from events import Event, EventSignup
 
 def get_domain():
     domain = {}
 
     # generate from model
-    for model in models.Base._decl_class_registry.values():
+    for model in Base._decl_class_registry.values():
         if not isinstance(model, DeclarativeMeta):
+            continue
+
+        # Exlude resources already put in modules
+        if model in [Event, EventSignup]:
             continue
 
         domain.update(make_domain(model))
@@ -25,8 +30,6 @@ def get_domain():
 
     # Only accept email addresses for email fields
     domain['users']['schema']['email'].update(
-        {'regex': EMAIL_REGEX})
-    domain['eventsignups']['schema']['email'].update(
         {'regex': EMAIL_REGEX})
 
     # /users
@@ -62,66 +65,6 @@ def get_domain():
 
     # POST will be handled by custom endpoint
     domain['sessions']['resource_methods'] = ['GET']
-
-    # /events
-
-    # additional validation
-    domain['events']['schema']['additional_fields'].update({
-        'type': 'json_schema'})
-    domain['events']['schema']['price'].update({'min': 0})
-    domain['events']['schema']['spots'].update({
-        'min': -1,
-        'if_this_then': ['time_register_start', 'time_register_end']})
-    domain['events']['schema']['time_register_end'].update({
-        'dependencies': ['time_register_start'],
-        'later_than': 'time_register_start'})
-    domain['events']['schema']['time_end'].update({
-        'dependencies': ['time_start'],
-        'later_than': 'time_start'})
-
-    # time_end for /events requires time_start
-    domain['events']['schema']['time_end'].update({
-        'dependencies': ['time_start']
-    })
-
-    # event images
-    domain['events']['schema'].update({
-        'img_thumbnail': {'type': 'media', 'filetype': ['png', 'jpeg']},
-        'img_banner': {'type': 'media', 'filetype': ['png', 'jpeg']},
-        'img_poster': {'type': 'media', 'filetype': ['png', 'jpeg']},
-        'img_infoscreen': {'type': 'media', 'filetype': ['png', 'jpeg']}
-    })
-
-    # /eventsignups
-
-    # POST done by custom endpoint
-    domain['eventsignups']['resource_methods'] = ['GET']
-
-    # schema extensions including custom validation
-    domain['eventsignups']['schema']['event_id'].update({
-        'not_patchable': True,
-        'unique_combination': ['user_id', 'email'],
-        'signup_requirements': True})
-    domain['eventsignups']['schema']['user_id'].update({
-        'not_patchable': True,
-        'unique_combination': ['event_id'],
-        'only_self_enrollment_for_event': True})
-    domain['eventsignups']['schema']['email'].update({
-        'not_patchable': True,
-        'unique_combination': ['event_id'],
-        'only_anonymous': True,
-        'email_signup_must_be_allowed': True})
-
-    # Since the data relation is not evaluated for posting, we need to remove
-    # it from the schema TODO: EXPLAIN BETTER
-    del(domain['eventsignups']['schema']['email']['data_relation'])
-    # Remove _email_unreg and _token from the schema since they are only
-    # for internal use and should not be visible
-    del(domain['eventsignups']['schema']['_email_unreg'])
-    del(domain['eventsignups']['schema']['_token'])
-
-    domain['eventsignups']['schema']['additional_fields'].update({
-        'type': 'json_event_field'})
 
     # /groups
 
@@ -176,5 +119,18 @@ def get_domain():
         'logo': {'type': 'media', 'filetype': ['png', 'jpeg']},
         'pdf': {'type': 'media', 'filetype': ['pdf']},
     })
+
+    # Add entry for storage so auth can find it
+
+    domain['storage'] = {
+        'resource_methods': ['GET'],
+        'item_methods': ['GET'],
+        'public_methods': [],
+        'public_item_methods': [],
+        'registered_methods': ['GET'],
+        'description': {
+            'general': 'Endpoint to download files, get the URLs via /files'
+        }
+    }
 
     return domain
